+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [0.0.1 - 2024-02-01]
+## [0.0.1 - 2024-02-02]
 ### Added
 - First version of the project
 - Spring Application
@@ -91,3 +91,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Check Expired Active Projects Job
 - Check if session scope is active before creating a notification for changing project state
 - Fetch token status and project status of token manager
+- New email if authentication script for DataSHIELD has changed